requirements.txt
tweepy==3.10.0
python-dotenv==0.15.0
APScheduler==3.7.0
tzlocal==2.1
selenium==3.141.0
